[hail] Rename `Table.scala` and move into `is.hail.expr.ir`
