+++ hail/src/main/scala/is/hail/expr/ir/Pretty.scala
@@ -3,7 +3,6 @@ package is.hail.expr.ir
 import is.hail.expr.JSONAnnotationImpex
 import is.hail.expr.ir.functions.RelationalFunctions
 import is.hail.expr.types.virtual.{TArray, TInterval}
-import is.hail.table.Ascending
 import is.hail.utils._
 import org.json4s.jackson.{JsonMethods, Serialization}
