+++ hail/src/main/scala/is/hail/expr/ir/PruneDeadFields.scala
@@ -3,7 +3,6 @@ package is.hail.expr.ir
 import is.hail.annotations._
 import is.hail.expr.types._
 import is.hail.expr.types.virtual._
-import is.hail.table.Ascending
 import is.hail.utils._
 
 import scala.collection.mutable
