+++ hail/src/main/scala/is/hail/expr/types/physical/PBaseStruct.scala
@@ -2,8 +2,7 @@ package is.hail.expr.types.physical
 
 import is.hail.annotations._
 import is.hail.asm4s.{Code, _}
-import is.hail.expr.ir.EmitMethodBuilder
-import is.hail.table.SortOrder
+import is.hail.expr.ir.{EmitMethodBuilder, SortOrder}
 import is.hail.utils._
 
 object PBaseStruct {
