+++ hail/src/main/scala/is/hail/expr/types/physical/PStruct.scala
@@ -2,10 +2,9 @@ package is.hail.expr.types.physical
 
 import is.hail.annotations._
 import is.hail.asm4s.Code
-import is.hail.expr.ir.EmitMethodBuilder
+import is.hail.expr.ir.{EmitMethodBuilder, SortOrder}
 import is.hail.expr.types.BaseStruct
 import is.hail.expr.types.virtual.{Field, TStruct, Type}
-import is.hail.table.SortOrder
 import is.hail.utils._
 import org.apache.spark.sql.Row
