--- hail/src/main/scala/is/hail/table/package.scala
@@ -1,7 +1,0 @@
-package is.hail
-
-package object table {
-  def asc(field: String): SortField = SortField(field, Ascending)
-
-  def desc(field: String): SortField = SortField(field, Descending)
-}
