+++ hail/src/test/scala/is/hail/expr/ir/OrderingSuite.scala
@@ -8,7 +8,6 @@ import is.hail.asm4s._
 import is.hail.TestUtils._
 import is.hail.expr.types.physical._
 import is.hail.expr.types.virtual._
-import is.hail.table.{ Ascending, Descending, SortOrder }
 import is.hail.utils._
 import org.apache.spark.sql.Row
 import org.testng.annotations.{DataProvider, Test}
