+++ hail/src/test/scala/is/hail/expr/ir/PruneSuite.scala
@@ -6,7 +6,6 @@ import is.hail.expr.types._
 import is.hail.expr.types.virtual._
 import is.hail.methods.{ForceCountMatrixTable, ForceCountTable}
 import is.hail.rvd.{RVD, RVDType}
-import is.hail.table._
 import is.hail.utils._
 import org.apache.spark.sql.Row
 import org.testng.annotations.{DataProvider, Test}
